pyproject.toml
[project]
name = "ai-chat-app"
version = "0.1.0"
dependencies = [
    "langchain-openai>=0.1.4",
    "streamlit==1.31.0",
    "langchain>=0.2.0",
    "ipykernel>=6.30.1",
    "fastapi>=0.136.0",
    "uvicorn>=0.44.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "langgraph>=1.1.8",
    "langgraph-checkpoint>=4.0.2",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "rich>=13.9.4",
    "duckduckgo-search>=8.1.1",
    "langchain-community>=0.4.1",
    "chromadb>=1.5.8",
    "sentence-transformers>=3.0.0",
    "onnxruntime>=1.17.0",
    "loguru>=0.7.3",
    "opentelemetry-api>=1.41.0",
    "opentelemetry-sdk>=1.41.0",
]
requires-python = ">=3.12, <3.13"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "W", "B", "SIM", "C4", "RUF"]
ignore = ["E501", "N818", "B018"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pyright>=1.1.409",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.11",
]
